import tushare misc data

+++ quantdatasource/jobs/tushare_misc_data.py
@@ -1,6 +1,7 @@
 import logging
 import pathlib
 
+import numpy as np
 import pandas as pd
 from quantcalendar import pydt_from_second
 
@@ -199,25 +200,29 @@ def tushare_misc_data(dt, is_collect, is_import):
             api.moneyflow_addition_path,
             chinese_names,
         )
-        stock_daily_out = output_dir / "bars_stock_daily"
+        stock_daily_out = output_dir / "daily_factors"
         stock_daily_out.mkdir(parents=True, exist_ok=True)
         stock_daily_file_path = stock_daily_out / f"{dt.date().isoformat()}.parquet"
         daily_bars.to_parquet(stock_daily_file_path)
         logging.info(f"写入[{stock_daily_file_path}]")
 
         yesterday = pydt_from_second(calendar.get_tradedays_lte(dt, 2)[0])
-        yesterday_daily_bars = pd.read_feather(
+        yesterday_daily_bars = pd.read_parquet(
             stock_daily_out / f"{yesterday.date().isoformat()}.parquet"
         )
         adjust_factors_collection = conn["finance"]["adjust_factors"]
         adj_factors = stock_utils.cal_adjust_factors(daily_bars, yesterday_daily_bars)
         for symbol, adj in adj_factors.items():
             lst = []
-            for one in adjust_factors_collection.find({"symbol": symbol}):
+            docs = list(adjust_factors_collection.find({"symbol": symbol}, projection={"_id": False}))
+            last_adj_date =docs[-1]["tradedate"]
+            if last_adj_date >= dt:
+                logging.warning(f"复权因子已经写入 {symbol}, last tradedate {last_adj_date}")
+                continue
+            for one in docs:
                 one["adjust_factor"] *= adj
-                print(one)
                 lst.append(one)
-            lst.append({"symbol": symbol, "adjust_factor": 1.0, "tradedate": dt})
+            lst.append({"symbol": symbol, "adjust_factor": np.float32(1.0), "tradedate": dt})
             adjust_factors_collection.delete_many({"symbol": symbol})
             adjust_factors_collection.insert_many(lst)
